Added signed() member function to intbv
The intbv class got a signed() member function added. The function will
classify the value of an intbv instance as unsigned or signed based on
the min/max settings. If a value is classified as unsigned, the bit
pattern is interpreted as a 2's-complement value and returned. If the
value is classified as signed, the value will be returned as is. The
signed() function will return the value as integer.

Added test cases verify the proper function of the signed() function.