Added index entry for enum() function usage
When using the index of the documentation I was missing a link for the
enum() usage. There is an entry for the enum() function that links to
the "Miscellaneous modelling support functions" section, but there is a
much better explanation of the enum() function in the "Finite State Machine
modeling" chapter. This changeset adds an index entry for enum() to this
chapter.