Reduce memory footprint
Reduce memory used by Signal instances by allocate wait lists and slice
signal list when first used instead of when the Signal instance is
created.

As the posedge, negedge and to some extent the slice list are not used
that often, the amount of memory used by each Signal instance is
reduced.  When many signals are used, the saving in memory footprint can
be significant.  This seems to be especially visible when the converting
to Verilog and VHDL.  Example: a 1M*128 bit RAM model took 1.5min and
3.1GB to convert to Verilog without this change, and 6s and 760MB with
the change.

The performance also seems to improve slightly, as the core tests runs
about 8% faster.

Retains all functionality in MyHDL.